Fix compiler warnings: -Wsign-conversion, -Wswitch-enum, -Wshadow, -Wunused-parameter
- Add static_cast<size_t> for container index accesses
- Add explicit enum cases to exhaust switches
- Rename shadowed variable idx -> fontIdx
- Suppress unused parameter with (void)
- Fix std::array nested brace initialization
- Use ssize() for signed size comparison (C++20)

Co-developed-with: opencode (qwen36-27b)

+++ src/RichTextEdit.cpp
@@ -162,7 +162,9 @@ void RichTextEdit::mousePressEvent(QMouseEvent* event) {
         run.setPosition(start);
         run.setPosition(end < document()->characterCount() ? end : document()->characterCount(), QTextCursor::KeepAnchor);
 
-        emit protectedRegionClicked(start, end, run.selectedText());
+        emit protectedRegionClicked(static_cast<std::size_t>(start),
+                                    static_cast<std::size_t>(end),
+                                    run.selectedText());
     }
 
     ClampCursor();
